+++ pkg/bitesize/environment_test.go
@@ -14,8 +14,8 @@ func TestExistingEnvironment(t *testing.T) {
 		t.Errorf("Unexpected error loading environment: %s", err.Error())
 	}
 
-	if len(e.Services) != 4 {
-		t.Errorf("Unexpected count of services. Expected 4, got: %d", len(e.Services))
+	if len(e.Services) != 6 {
+		t.Errorf("Unexpected count of services. Expected 6, got: %d", len(e.Services))
 	}
 
 }