+++ pkg/diff/compare.go
@@ -29,13 +29,16 @@ func Compare(config1, config2 bitesize.Environment) string {
 				s.Version = d.Version
 			}
 			s.Status = d.Status
-			s.Deployment = d.Deployment
 		}
+		// Ignore diff for annotations if d.Deployment is nil
+		if d == nil || d.Deployment == nil {
+			s.Annotations = nil
+		}
+
 		newServices = append(newServices, s)
 		// }
 	}
 	c1.Services = newServices
-	// XXX: the end
 
 	// Ignore diff between service replicas when hpa is configured
 	var correctedServices bitesize.Services